+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 1.1.0 - 2017-1-2
 
+- Fixed overwrite save function `Cannot read property 'destroy' of null`
 - Added `count`, `findOne` & `findAll` functions
 
 ## 1.0.2 - 2017-1-1